Code Push 30 April 2021
Authentication added to the admin page. Removed unused sections to simplify app. Style added to the login page so it looks good. Added footer section - still need to figure out why the footer is in the middle and not extended  (should be responsive, but this part is not). In terms of completeness, I would say we are at 75%